Moved IndexTemplate and RebuildableIndex into same fulltextsearch package as the rest
