+++ whois-api/src/main/java/net/ripe/db/whois/api/fulltextsearch/FullTextIndex.java
@@ -4,8 +4,6 @@ import com.google.common.base.CharMatcher;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import net.ripe.db.whois.api.search.IndexTemplate;
-import net.ripe.db.whois.api.search.RebuildableIndex;
 import net.ripe.db.whois.common.dao.jdbc.JdbcRpslObjectOperations;
 import net.ripe.db.whois.common.dao.jdbc.JdbcStreamingHelper;
 import net.ripe.db.whois.common.domain.serials.SerialEntry;
